gsuite-cli/requirements.txt
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
google-api-core==1.19.1
google-api-python-client==1.9.1
google-auth==1.16.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.9
oauthlib==3.1.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
random-password-generator==2.1.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.9
